✅ test(home): use kanji cards for JLPT progress E2E test
- Revert <1% display over-engineering back to simple {:.0}%
- Move JLPT progress test from lesson.spec.ts to home.spec.ts
- Use KanjiPage (N5 kanji, ~80 items) instead of WordsPage (616 items)
- 5 marked kanji gives ~6% kanji progress → visible in overall %
